HW: Replace Server/Client classes with Instance, add Input
The Server/Client classes have been aggregated to consider the format of
the Processing Network class, which handles events using handler
functions.

The Input class is a special subclass of the Display class for use in
instances for command/message input.

Due to the reduction of diagram complexity, the UML diagram is now
vertically oriented.